Update Makefile and .ignore files (#1031)
This commit updated the Makefile and .ignore files for the evaluator and mmf binaries.

Also moved the evaluator to test/evaluator folder - I had it accidentally placed under the test/customize/evaluator dir because of a bad merge when working on deprecating the harness.
