--- test/customize/matchfunction/Dockerfile
@@ -1,24 +1,0 @@
-# Copyright 2019 Google LLC
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-FROM open-match-base-build as builder
-
-WORKDIR /go/src/open-match.dev/open-match/test/customize/matchfunction
-RUN CGO_ENABLED=0 GOOS=linux go build -a -installsuffix cgo -o matchfunction .
-
-FROM gcr.io/distroless/static:nonroot
-WORKDIR /app/
-COPY --from=builder --chown=nonroot /go/src/open-match.dev/open-match/test/customize/matchfunction/matchfunction /app/
-
-ENTRYPOINT ["/app/matchfunction"]
